+++ ChangeLog.md
@@ -1,3 +1,7 @@
+## Changes between o3d3xx-h5 0.1.1 and 0.1.2
+
+* Aware of the `raw_amplitude` image
+
 ## Changes between o3d3xx-h5 0.1.0 and 0.1.1
 
 * Aware of the change from `xyzi_image` to `xyz_image` in o3d3xx-ros